refactor: move signIn auth function
