+++ src/shared/buttons/ButtonTFMain/ButtonTFMain.tsx
@@ -6,6 +6,7 @@ type Props = {
 const ButtonTFMain = ({ onClick, label }: Props) => {
   return (
     <button
+      type="submit"
       onClick={onClick}
       className="text-p text-text-color leading-8 cursor-pointer bg-button-primary py-13 px-30 rounded-100 hover:shadow-hover-btn hover:bg-hover-btn-primary transition-all w-min"
     >
